Updated Login to support FakeData and login service
What I did:
1.	Added FakeData class in front end shared folder
a.	Contains a get for authotekn generated in Authoken.ts
b.	Contains a list of Profiles
c.	Contains a function to get the first Proflile in the list to use for login
2.	Added Authtoken class in front end shared folder
a.	Generates an authoken either via uuid or random string generation
3.	Added LoginService class in services folder in src of front end
a.	Contains the function login that grabs the first Profile from FakeData and returns it.
4.	Added UserInfoProvider class in components folder in src of front end
a.	Class used to globally and locally store user information to be used across the different pages/views after logging in or registering.
5.	Added < UserInfoProvider > to App.tsx
6.	Updated Login.tsx to call login service and update user info in UserInfoProvider
7.	Updated Login.css to fit the prototype