Updated Register Page
I did my best to throw together the Register page as quickly as I could. Some of the responsiveness is a little rough and the code can get a little repetitive. But here’s what I did
1.	Create 8 different steps that lead the user through the registration process like in the prototype.
2.	Collected information is listed under the User info state on the Register.tsx file but for simplicity its: owner first name, owner last name, dog name, email, password, owners birthday, dogs birthday, street, city, zip code, dog breed, list of dog traits, and a photo of the dog.
3.	Set up a handle function for register
PLEASE NOTE, I saw that the register service right now just seems to be connecting to the fake data so I didn’t touch it, as I wasn’t sure what you might want to do with it. I ONLY provided the handle function to call it in step 8 (the last step) in the Register.tsx file when the user clicks the continue button in that step.
Hope this page set up works for you, I’ve never used step states or step components before.